requirements.txt
selenium==4.40.0
pillow==12.1.0
webdriver-manager==4.0.2
requests>=2.28.0
numpy>=1.24.0